backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.35
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
ortools==9.11.4210
openai==1.51.2
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
ruff>=0.1.0
mypy>=1.0.0
